prohlizecka, export do zip funkcni, stav pred odeslanim Romanovi

+++ app/src/main/java/com/tomst/lolly/core/OnProListener.java
@@ -2,5 +2,5 @@ package com.tomst.lolly.core;
 
 @FunctionalInterface
 public interface OnProListener {
-    void OnProEvent(long Pos);
+    void OnProEvent(int Pos);
 }
